Fixed compiler warnings.

+++ src/gallium/state_trackers/clover/core/platform.cpp
@@ -135,7 +135,7 @@ platform::load_amdocl2() {
    amdocl2_dynlib.load(amdocl2_cur_path.c_str(), DYNLIB_NOW);
    
    amdocl2_funcs.reset(new amdocl2_funcs_struct);
-   for(int i = 0; i < sizeof(amdocl2_funcs_struct)/sizeof(void*); i++)
+   for(size_t i = 0; i < sizeof(amdocl2_funcs_struct)/sizeof(void*); i++)
       amdocl2_funcs->funcs[i] = amdocl2_dynlib.getSymbol(amdocl2_funcs_names[i]);
    
    cl_platform_id platform;
